Move LAN bootstrap orchestration into network
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
