+++ crates/mesh-llm-host-runtime/src/network/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod affinity;
 pub(crate) mod discovery;
 pub(crate) mod lan_beacon;
+pub(crate) mod lan_bootstrap;
 pub(crate) mod mdns_reverse_dial;
 pub(crate) mod metrics;
 pub(crate) mod nostr;
